// src/app/models/ActionCardTitles.ts
export enum ActionCardTitles {
    DealBreaker = "Deal Breaker",
    JustSayNo = "Just Say No",
    SlyDeal = "Sly Deal",
    ForcedDeal = "Forced Deal",
    DebtCollector = "Debt Collector",
    ItsMyBirthday = "Its My Birthday",
    PassGo = "Pass Go",
    House = "House",
    Hotel = "Hotel",
    DoubleTheRent = "Double The Rent"
}